Migrate to store builder
