+++ react/ProductWrapper.js
@@ -16,6 +16,10 @@ class ProductWrapper extends Component {
     query: PropTypes.object,
   }
 
+  stripCategory(category) {
+    return category && category.replace(/^\/|\/$/g, '')
+  }
+
   getData = () => {
     const {
       productQuery: { product },
