--- react/Queries.js
@@ -1,13 +1,0 @@
-import session from './queries/sessionQuery.gql'
-import orderForm from './queries/orderFormQuery.gql'
-import product from './queries/productQuery.gql'
-import recommendatinsAndBenefits from './queries/recommendationsAndBenefitsQuery.gql'
-import search from './queries/searchQuery.gql'
-
-export default {
-  orderForm,
-  product,
-  recommendatinsAndBenefits,
-  search,
-  session,
-}
