--- react/queries/orderFormQuery.gql
@@ -1,57 +1,0 @@
-query orderForm {
-  orderForm {
-    cacheId
-    orderFormId
-    value
-    totalizers {
-      id
-      name
-      value
-    }
-    items {
-      id
-      name
-      imageUrl
-      detailUrl
-      skuName
-      quantity
-      sellingPrice
-      listPrice
-    }
-    shippingData {
-      address {
-        id
-        neighborhood
-        complement
-        number
-        street
-        postalCode
-        city
-        reference
-        addressName
-        addressType
-      }
-      availableAddresses {
-        id
-        neighborhood
-        complement
-        number
-        street
-        postalCode
-        city
-        reference
-        addressName
-        addressType
-      }
-    }
-    clientProfileData {
-      email
-      firstName
-    }
-    storePreferencesData {
-      countryCode
-      currencyCode
-      timeZone
-    }
-  }
-}
