--- react/queries/productQuery.gql
@@ -1,136 +1,0 @@
-query Product($slug: String) {
-  product(slug: $slug) {
-    cacheId
-    productName
-    productId
-    description
-    titleTag
-    metaTagDescription
-    linkText
-    categories
-    categoryId
-    categoriesIds
-    brand
-    properties {
-      name
-      values
-    }
-    items {
-      itemId
-      name
-      nameComplete
-      complementName
-      ean
-      referenceId {
-        Key
-        Value
-      }
-      measurementUnit
-      unitMultiplier
-      images (quantity: 10) {
-        imageId
-        imageLabel
-        imageTag
-        imageUrl
-        imageText
-      }
-      sellers {
-        sellerId
-        sellerName
-        addToCartLink
-        sellerDefault
-        commertialOffer {
-          Price
-          ListPrice
-          PriceWithoutDiscount
-          RewardValue
-          PriceValidUntil
-          AvailableQuantity
-          Tax
-          CacheVersionUsedToCallCheckout
-          Installments {
-              Value
-              InterestRate
-              TotalValuePlusInterestRate
-              NumberOfInstallments
-              Name
-          }
-        }
-      }
-      variations {
-        name
-        values
-      }
-      kitItems {
-        itemId
-        amount
-        product {
-          productName
-          productId
-    			description
-    			linkText
-    			categories
-    			categoryId
-    			categoriesIds
-    			brand
-    			properties {
-      			name
-      			values
-    			}
-        }
-        sku {
-          itemId
-          name
-          referenceId {
-            Key
-            Value
-          }
-          images {
-            imageId
-            imageLabel
-            imageTag
-            imageUrl
-            imageText
-          }
-          sellers {
-            sellerId
-            sellerName
-            addToCartLink
-            sellerDefault
-            commertialOffer {
-              Price
-              ListPrice
-              PriceWithoutDiscount
-              RewardValue
-              PriceValidUntil
-              AvailableQuantity
-              Tax
-              CacheVersionUsedToCallCheckout
-              Installments {
-                  Value
-                  InterestRate
-                  TotalValuePlusInterestRate
-                  NumberOfInstallments
-                  Name
-              }
-            }
-          }
-        }
-      }
-      attachments {
-        id
-        name
-        required
-      }
-      calculatedAttachments
-    }
-    productClusters{
-      id
-      name
-    }
-    clusterHighlights{
-      id
-      name
-    }
-  }
-}
