--- react/queries/searchQuery.gql
@@ -1,97 +1,0 @@
-query search($query: String, $map: String, $rest: String, $orderBy: String, $priceRange: String, $from: Int, $to: Int, $withFacets: Boolean = true) {
-  search(query: $query, map: $map, rest: $rest, orderBy: $orderBy, priceRange: $priceRange, from: $from, to: $to) {
-    titleTag
-    metaTagDescription
-    facets @include (if: $withFacets) {
-      Departments {
-        Quantity
-        Name
-        Link
-      }
-      Brands {
-        Quantity
-        Name
-        Link
-      }
-      SpecificationFilters {
-        name
-        facets {
-          Quantity
-          Name
-          Link
-        }
-      }
-      CategoriesTrees {
-        Id
-        Quantity
-        Name
-        Link
-        Children {
-          Id
-          Quantity
-          Name
-          Link
-          Children {
-            Id
-            Quantity
-            Name
-            Link
-          }
-        }
-      }
-      PriceRanges {
-        Quantity
-        Name
-        Link
-        Slug
-      }
-    }
-    products {
-      cacheId
-      categories
-      productId
-      productName
-      linkText
-      description
-      brand
-      link
-      items {
-        itemId
-        name
-        nameComplete
-        complementName
-        ean
-        referenceId {
-          Key
-          Value
-        }
-        measurementUnit
-        unitMultiplier
-        images {
-          cacheId
-          imageId
-          imageLabel
-          imageTag
-          imageUrl
-          imageText
-        }
-        sellers {
-          sellerId
-          sellerName
-          commertialOffer {
-            Price
-            ListPrice
-            PriceWithoutDiscount
-            RewardValue
-            PriceValidUntil
-            AvailableQuantity
-          }
-        }
-      }
-      productClusters {
-        name
-      }
-    }
-    recordsFiltered
-  }
-}
