--- react/queries/sessionQuery.gql
@@ -1,23 +1,0 @@
-query Session {
-  getSession {
-    id
-    adminUserId
-    impersonable
-    adminUserEmail
-    profile {
-      id
-      firstName
-      lastName
-      email
-    }
-    impersonate {
-      profile {
-        phone
-        email
-        lastName
-        document
-        firstName
-      }
-    }
-  }
-}
